refactor(ontology): flatten ModelInfo into ModelProvenance, remove model module
- Move ModelKind into provenance.rs alongside ModelProvenance
- Flatten ModelInfo fields (name, kind, version) directly onto
  ModelProvenance — removes the unnecessary wrapper struct
- Delete entity/model/ module entirely
- RecognitionMethod::ner/classification/biometric now take
  (name, ModelKind) directly instead of ModelInfo
- Update all consumers across engine, pattern, and provider crates
